Improve OpenAPI definition
I am one of the developers of the [Micronaut Framework](https://micronaut.io/). As part of writing a guide for our [OpenAPI support](https://github.com/micronaut-projects/micronaut-openapi/), I was contemplating using Open Meteo as an example service.

I was happy to find it was coming with an OpenAPI spec. However, the response types are not precise enough for our generated code. Typically, return types are often `Object` instead of something strongly typed.

This commit fixes this by introducing component types with the appropriate return types. I'm not an OpenAPI expert but I have verified that this fixes the issue I was seeing.